// src/API.ts
export type CreatePostInput = {
  id?: string | null,
  title: string,
  content: string,
  image?: string | null,
  upvotes?: number | null,
  downvotes?: number | null,
};
// ...
export type Post = {
  __typename: "Post",
  id: string,
  title: string,
  content: string,
  image?: string | null,
  upvotes?: number | null,
  downvotes?: number | null,
  createdAt: string,
  updatedAt: string,
  owner?: string | null,
  comments?: ModelCommentConnection | null,
};

export type ModelCommentConnection = {
  __typename: "ModelCommentConnection",
  items?:  Array<Comment | null > | null,
  nextToken?: string | null,
};

export type Comment = {
  __typename: "Comment",
  id: string,
  postID: string,
  content: string,
  createdAt: string,
  updatedAt: string,
  post?: Post | null,
  owner?: string | null,
};
// ...
export type CreatePostMutation = {
  createPost?:  {
    __typename: "Post",
    id: string,
    title: string,
    content: string,
    image?: string | null,
    upvotes?: number | null,
    downvotes?: number | null,
    createdAt: string,
    updatedAt: string,
    owner?: string | null,
    comments?:  {
      __typename: "ModelCommentConnection",
      items?:  Array< {
        __typename: "Comment",
        id: string,
        postID: string,
        content: string,
        createdAt: string,
        updatedAt: string,
        owner?: string | null,
      } | null > | null,
      nextToken?: string | null,
    } | null,
  } | null,
};